backend/requirements.txt
Flask==2.2.3
Werkzeug==2.3.8
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.0
pymysql==1.1.1
cryptography==41.0.1
neo4j==5.20.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
numpy==1.26.4
pandas==2.0.2
requests==2.31.0
python-dateutil==2.8.2
bcrypt==4.0.1
pydantic==1.10.8
faiss-cpu==1.7.4
scikit-learn==1.3.0
matplotlib==3.7.2
gunicorn==21.2.0
